created FavoriteCocktail component for favorite button, added hover effect on buttons, added skiplink, styled savedcocktails
