+++ public/index.html
@@ -31,6 +31,7 @@
     <title>React App</title>
   </head>
   <body>
+    <a href="#mainContent" class="skipLink">Skip to the main content.</a>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
@@ -42,6 +43,6 @@
 
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
-    -->
+    -->    
   </body>
 </html>
