+++ src/App.js
@@ -1,15 +1,13 @@
 import './App.css';
 import Header from './components/Header';
 import  Cocktail from './components/Cocktail';
-import FavoriteCocktail from './components/FavoriteCocktail';
 
 
 function App() {
   return (
     <>
       <Header />
-      <Cocktail />
-      <FavoriteCocktail />
+      <Cocktail />     
     </>
   );
 }
