+++ src/components/Cocktail.js
@@ -1,6 +1,7 @@
-import CocktailItem from './CocktailItem';
 import { useState } from 'react';
 import axios from 'axios';
+import CocktailItem from './CocktailItem';
+import FavoriteCocktail from './FavoriteCocktail';
 
 const Cocktail = () => {
 
@@ -11,7 +12,6 @@ const Cocktail = () => {
         url: "https://thecocktaildb.com/api/json/v1/1/random.php"  
         })    
         .then( (data) => {
-            console.log(data)
             setCocktailData(data.data.drinks[0])
         })
     }
@@ -20,6 +20,10 @@ const Cocktail = () => {
         <>
             <CocktailItem cocktailData={cocktailData}/>
             <button className="randomButton" onClick= { handleClick } >Surprise Me</button>
+            <footer>
+                <a className="footer" href="https://junocollege.com/">Created at Juno</a>
+            </footer>            
+            <FavoriteCocktail cocktailData={cocktailData}/>                  
         </>
     )
 }
